CSS + Edit Student

+++ final-project/delete.php
@@ -1,9 +1,9 @@
 <?php
 include 'db.php';
 
-$id = $_GET['id'];
-
-$conn->query("DELETE FROM students WHERE id=$id");
-
+$id = (int) ($_GET['id'] ?? 0);
+if ($id >= 1) {
+    $conn->query("DELETE FROM students WHERE id=$id");
+}
 header("Location: index.php");
-?>
+exit;
